feat: implement bv_normalize, A bitvector normalizer for multiplication terms
For example, `x₁ * (y₁ * z) = x₂ * (y₂ * z)` is normalized to
`z * (x₁ * y₁) = z * (x₂ * y₂)`,
pulling the shared variable `z` to the front on both sides.
The PR also replaces the use of `ac_nf` in the normalization pass of `bv_decide`.

Note that this is based on Bitwuzla's normalizer,
and we eventually want to have support for bitvector addition normalization as well.
However, since we currently lack a `ring` equivalent for bitvectors,
we cannot currently justify rewrites such as `x + x + x → 3 * x`.

We only run the normalization on terms of the form `(a * b)  = (c * d)` for `a, b, c, d` bitvectors.
This mirrors Bitwuzla's `PassNormalize::process`'s `PassNormalize::normalize_eq_add_mul`.
We leave the implementation of `PassNormalize::normalize_comm_assoc`,
which is called when the toplevel terms are different for a subsequent patch.

For posterity, we record the precise location in Bitwuzla where the codepath occurs:
```cpp
-- https://github.com/bitwuzla/bitwuzla/blob/d1f1bc2ad3963936d3167afb879867364554cee7/src/preprocess/pass/normalize.cpp#L1550-L1554
        Kind k = cur.kind();
        if (k == Kind::EQUAL && children[0].kind() == children[1].kind()
            && (children[0].kind() == Kind::BV_ADD
                || children[0].kind() == Kind::BV_MUL))
        {
          auto [res, norm] = normalize_eq_add_mul(children[0], children[1]);
          ...
```