feat: kernel: lazy_delta_reduction: try to reduce_nat
This PR elaborates on #7386 (which was merged as part of #7387), and
allows `lazy_delta_reduction` to call `reduce_nat` also on open terms.

Without this we get
```
/--
error: (kernel) declaration type mismatch, '_example' has type
  ∀ (n : Nat), Function.const Nat 0 n = Function.const Nat 0 n
but it is expected to have type
  ∀ (n : Nat), Function.const Nat 0 n = Nat.div 0 2
-/
example (n : Nat) :
  Function.const Nat (0 : Nat) n = Nat.div 0 2 := rfl
```
once the well-founded proof passsed to `fix` is irreducible.