fix: consistent term order in linear integer normalization
This PR ensures that we use the same ordering to normalize linear
`Int` terms and relations. This change affects `simp +arith` and
`grind` normalizer.

This consistency is important in the cutsat procedure.
We want to avoid a situation where the cutsat state contains both
"atoms":
- `「(NatCast.natCast x + NatCast.natCast y) % 8」`
- `「(NatCast.natCast y + NatCast.natCast x) % 8」`

This  was happening because we were using different orderings for
(nested) terms and relations (`=`, `<=`).