new induction syntax
feat: allow omission of `=> ?_` in `induction`/`cases`

This PR modifies the syntax of `induction`, `cases`, and other tactics that use `Lean.Parser.Tactic.inductionAlts`. If a case omits `=> ...` then it is assumed to be `=> ?_`. For example,
```
example (p : Nat × Nat) : p.1 = p.1 := by
  cases p with | _ p1 p2
  /-
  case mk
  p1 p2 : Nat
  ⊢ (p1, p2).fst = (p1, p2).fst
  -/
```
This works with multiple cases as well. For example, `induction n with | zero | succ n ih` is short for `induction n with | zero | succ n ih => ?_` is short for `induction n with | zero => ?_ | succ n ih => ?_`. Note that the grammar makes it so that only the last alternative can omit `=>`. Any `=>`-free alternatives before an alternative with `=>` will take that `=>`.

fix macro expansion loop